Moving Ticket.java and TicketService into webAppBackend and deleting Main.java - this resolves #51
