+++ backend/webAppBackend/webAppBackend/pom.xml
@@ -51,6 +51,17 @@
 			<artifactId>google-cloud-firestore</artifactId>
 			<version>3.26.0</version>
 		</dependency>
+
+		<dependency>
+			<groupId>com.google.zxing</groupId>
+			<artifactId>core</artifactId>
+			<version>3.5.3</version> <!-- Use the latest stable version -->
+		</dependency>
+		<dependency>
+			<groupId>com.google.zxing</groupId>
+			<artifactId>javase</artifactId>
+			<version>3.5.3</version> <!-- Use the latest stable version -->
+		</dependency>
 	</dependencies>
 
 	<build>
